+++ src/Ceres.Chess/LC0/Positions/Batches/IEncodedPositionBatchFlat.cs
@@ -108,7 +108,7 @@ namespace Ceres.Chess.LC0.Batches
 
     #region Implmentation
 
-    float[] ValuesFlatFromPlanes(float[] preallocatedBuffer, bool nhwc, bool scale50MoveCounter);
+    Half[] ValuesFlatFromPlanes(Half[] preallocatedBuffer, bool nhwc, bool scale50MoveCounter);
 
     
     public IEncodedPositionBatchFlat GetSubBatchSlice(int startIndex, int count)
